As a part of removing the AppLauncherService implementation.
* Removed unnecessary properties.
* Removed unnecessary tests.
* Removed methods from WorkbenchDataManager without use.

Issue: IBP-4878
Review: none
