+++ src/main/java/org/generationcp/middleware/api/program/ProgramServiceImpl.java
@@ -97,4 +97,8 @@ public class ProgramServiceImpl implements ProgramService {
 		this.daoFactory.getUserRoleDao().removeUsersFromProgram(userIds, projectId);
 	}
 
+	@Override
+	public long countProjectsByFilter(final ProgramSearchRequest programSearchRequest) {
+		return this.daoFactory.getProjectDAO().countProjectsByFilter(programSearchRequest);
+	}
 }
