+++ src/test/java/org/generationcp/middleware/api/program/ProgramServiceImplIntegrationTest.java
@@ -7,6 +7,7 @@ import org.generationcp.middleware.manager.WorkbenchDaoFactory;
 import org.generationcp.middleware.pojos.workbench.CropType;
 import org.generationcp.middleware.pojos.workbench.Project;
 import org.generationcp.middleware.pojos.workbench.WorkbenchUser;
+import org.generationcp.middleware.service.api.program.ProgramSearchRequest;
 import org.junit.Before;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,6 +15,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import java.util.Date;
 import java.util.UUID;
 
+import static org.hamcrest.CoreMatchers.equalTo;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.assertEquals;
 
 public class ProgramServiceImplIntegrationTest extends IntegrationTestBase {
@@ -48,6 +52,22 @@ public class ProgramServiceImplIntegrationTest extends IntegrationTestBase {
 		assertEquals(programDTO.getUniqueID(), testProject1.getUniqueID());
 	}
 
+	@Test
+	public void test_CountProjectsByFilter_Ok() {
+		final Integer userId = this.findAdminUser();
+		this.workbenchUserTest = this.daoFactory.getWorkbenchUserDAO().getById(userId);
+		this.sessionProvder.getSession().flush();
+
+		final ProgramSearchRequest programSearchRequest = new ProgramSearchRequest();
+		final Project project = buildProject("Project1 ");
+		programSearchRequest.setProgramName(project.getProjectName());
+		programSearchRequest.setCommonCropName(project.getCropType().getCropName());
+		programSearchRequest.setLoggedInUserId(userId);
+
+		final long count = this.programService.countProjectsByFilter(programSearchRequest);
+		assertThat(new Long(1), is(equalTo(count)));
+	}
+
 	private Project buildProject(final String projectName) {
 		final Project project = new Project();
 		project.setProjectName(projectName + RandomStringUtils.randomAlphanumeric(10));
@@ -59,4 +79,5 @@ public class ProgramServiceImplIntegrationTest extends IntegrationTestBase {
 		this.sessionProvder.getSession().flush();
 		return project;
 	}
+
 }
